create Summary component
