+++ src/pages/Account/Summary/testIds.ts
@@ -1,0 +1,7 @@
+const TEST_IDS = {
+  container: 'account-summary-container',
+  stats: 'account-summary-stats',
+  accountHealth: 'account-summary-account-health',
+};
+
+export default TEST_IDS;
