--- src/pages/Account/testIds.ts
@@ -1,5 +1,0 @@
-const TEST_IDS = {
-  stats: 'account-stats',
-};
-
-export default TEST_IDS;
